# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance and productivity API",
        "title": "Kaizen API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@kaizen.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/categories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all categories for the current user with optional type filter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Categories"
                ],
                "summary": "List finance categories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by type (income or expense)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new income or expense category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Categories"
                ],
                "summary": "Create a finance category",
                "parameters": [
                    {
                        "description": "Category details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single category by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Categories"
                ],
                "summary": "Get a finance category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Categories"
                ],
                "summary": "Update a finance category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Category data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a category (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Finance Categories"
                ],
                "summary": "Delete a finance category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jedi116_kaizen-api_internal_models.FinanceCategory": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "Hex color for UI (e.g., \"#FF5733\")",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "Optional description",
                    "type": "string"
                },
                "icon": {
                    "description": "Icon name or emoji (e.g., \"🍔\", \"💰\")",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "description": "Soft disable without deleting",
                    "type": "boolean"
                },
                "journals": {
                    "description": "Has many journals",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceJournal"
                    }
                },
                "name": {
                    "description": "e.g., \"Groceries\", \"Salary\", \"Entertainment\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"income\" or \"expense\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Which user owns this category",
                    "type": "integer"
                }
            }
        },
        "github_com_jedi116_kaizen-api_internal_models.FinanceJournal": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Transaction amount (always positive)",
                    "type": "number"
                },
                "category": {
                    "description": "Belongs to a category",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_jedi116_kaizen-api_internal_models.FinanceCategory"
                        }
                    ]
                },
                "category_id": {
                    "description": "Which category (Food, Rent, etc.)",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "description": "Transaction date (for daily tracking)",
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "description": "Optional detailed notes",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_recurring": {
                    "description": "Is this a recurring transaction?",
                    "type": "boolean"
                },
                "location": {
                    "description": "Where transaction occurred (optional)",
                    "type": "string"
                },
                "payment_method": {
                    "description": "\"cash\", \"credit_card\", \"bank_transfer\", etc.",
                    "type": "string"
                },
                "receipt_url": {
                    "description": "Attachments (optional - for receipts)",
                    "type": "string"
                },
                "title": {
                    "description": "e.g., \"Grocery shopping at Walmart\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"income\" or \"expense\"",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "description": "Which user made this transaction",
                    "type": "integer"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers.CreateCategoryRequest": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "color": {
                    "type": "string",
                    "example": "#FF5733"
                },
                "description": {
                    "type": "string",
                    "example": "Food and household items"
                },
                "icon": {
                    "type": "string",
                    "example": "🛒"
                },
                "name": {
                    "type": "string",
                    "example": "Groceries"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ],
                    "example": "expense"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "internal_handlers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "internal_handlers.UpdateCategoryRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string",
                    "example": "#FF5733"
                },
                "description": {
                    "type": "string",
                    "example": "Food and household items"
                },
                "icon": {
                    "type": "string",
                    "example": "🛒"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Groceries"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ],
                    "example": "expense"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}